Undo sqlalchemy changes
